refactor: move test folder outside src
