+++ jest.config.ts
@@ -1,5 +1,5 @@
 export default {
-  roots: ["<rootDir>/src"],
+  roots: ["<rootDir>/tests"],
   collectCoverageFrom: ["<rootDir>/src/**/*.ts"],
   coverageDirectory: "coverage",
   coverageProvider: "v8",
