+++ tests/generator.test.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from "fs";
 import { GraphQLSchema } from "graphql";
 import path from "path";
-import { getMocks } from "../generator";
+import { getMocks } from "../src/generator";
 
 const schema = readFileSync(
   path.join(__dirname, "./resources/testSchema.graphql"),
